raop: Set precise wake up timer for rtpoll
Due to timer precision limitation, poll(2) may sleep longer or shorter
than originally specified. Therefore it is not safe to always specify
the fixed sleep interval for poll(2) (pa_rtpoll_set_timer_relative())
when executing some periodic tasks that are sensitive to precise timing.

This patch calculates when the thread should wake up to transmit
the next audio packet, and try to amortize the jitter caused by the
actual poll sleep duration.

This patch eliminates the audio glitch issue in Pioneer VSX-43 (and
possibly in other devices as well).