raop: Set poll timer when entering the RUNNING state
When pausing music, the raop module enters the IDLE state.
If you resume the music right after pausing, the module enters
the RUNNING state but the poll timer is left expired.
This prevents the main thread loop from sending audio packets.

To avoid this situation, set the poll timer again when the module
enters the RUNNING state.