raop: Cleanup pollfd when the RTSP state enters TEARDOWN
It seems that STATE_DISCONNECTED never happens in usual cases,
which left pollfd uncleaned even after the socket fds were closed in
the previous version. This led to PulseAudio hang up when starting music
after the previous connection is torn down.

This patch fixed the issue, thus fixes the hang up issue in such a case.