raop: Add the core implementation of RAOP authentication
RAOP authentication (password) is based on BA and DA HTTP authentication
schemes. This patch adds the RSTP client the ability to specify the caller
of server response status. Tracking the '401 Unauthorized' status allow
the RAOP client to respond the server challenge authentication request.
This patch adds the core implementation but does not introduce the
authentication scheme in the RAOP connection process yet.